+++ vhost/src/vhost_kern/net.rs
@@ -68,7 +68,7 @@ impl<AS: GuestAddressSpace> AsRawFd for Net<AS> {
 
 #[cfg(test)]
 mod tests {
-    use vm_memory::{GuestAddress, GuestMemory, GuestMemoryMmap};
+    use vm_memory::{GuestAddress, GuestMemoryBackend, GuestMemoryMmap};
     use vmm_sys_util::eventfd::EventFd;
 
     use super::*;